requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==2.4.0
djangorestframework==3.10.3
gunicorn==20.0.0
haversine==2.1.2
idna==2.8
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
ufo==0.2.0
urllib3==1.25.7
